Develop talent

See merge request talentmap/talentcard!666

+++ talentweb/src/main/java/com/talentcard/web/service/impl/BannerServiceImpl.java
@@ -10,13 +10,16 @@ import com.talentcard.common.utils.FileUtil;
 import com.talentcard.common.utils.redis.RedisMapUtil;
 import com.talentcard.common.vo.PageInfoVO;
 import com.talentcard.common.vo.ResultVO;
+import com.talentcard.web.constant.OpsRecordMenuConstant;
 import com.talentcard.web.dto.BannerDTO;
 import com.talentcard.web.service.IBannerService;
+import com.talentcard.web.service.ILogService;
 import com.talentcard.web.vo.BannerVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.servlet.http.HttpSession;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
@@ -35,7 +38,8 @@ public class BannerServiceImpl implements IBannerService {
     private FilePathConfig filePathConfig;
     @Autowired
     private RedisMapUtil redisMapUtil;
-
+    @Autowired
+    private ILogService logService;
     @Override
     public ResultVO query(int pageNum, int pageSize, Map<String, Object> reqMap) {
         Page<BannerPO> page = PageHelper.startPage(pageNum, pageSize);
@@ -44,7 +48,13 @@ public class BannerServiceImpl implements IBannerService {
     }
 
     @Override
-    public ResultVO insert(BannerDTO dto) {
+    public ResultVO insert(HttpSession session,BannerDTO dto) {
+        //从session中获取userId的值
+        Long userId = (Long) session.getAttribute("userId");
+        if (userId == null) {
+            // 用户过期
+            return ResultVO.notLogin();
+        }
         BannerPO existPO = bannerMapper.queryByName(dto.getName());
         if(null != existPO) {
             return new ResultVO(1103);
@@ -54,11 +64,19 @@ public class BannerServiceImpl implements IBannerService {
         po.setStatus((byte) 2);
         po.setDr((byte) 1);
         bannerMapper.insert(po);
+        logService.insertActionRecord(session, OpsRecordMenuConstant.F_ExternalFunction,OpsRecordMenuConstant.S_BannerConfig,
+                "新建banner\"%s\"",po.getName());
         return new ResultVO(1000);
     }
 
     @Override
-    public ResultVO status(Long bid, Byte status) {
+    public ResultVO status(HttpSession session,Long bid, Byte status) {
+        //从session中获取userId的值
+        Long userId = (Long) session.getAttribute("userId");
+        if (userId == null) {
+            // 用户过期
+            return ResultVO.notLogin();
+        }
         BannerPO po = bannerMapper.selectByPrimaryKey(bid);
         if (null == po) {
             return new ResultVO(1101);
@@ -73,19 +91,37 @@ public class BannerServiceImpl implements IBannerService {
         bannerMapper.updateByPrimaryKey(po);
 
         this.redisMapUtil.del(TalentConstant.BANNER_INFO);
-
+        logService.insertActionRecord(session,OpsRecordMenuConstant.F_ExternalFunction,OpsRecordMenuConstant.S_BannerConfig,
+                "%s架banner\"%s\"",1 == status?"上":"下",po.getName());
         return new ResultVO(1000);
     }
 
     @Override
-    public ResultVO delete(Long bid) {
+    public ResultVO delete(HttpSession session,Long bid) {
+        //从session中获取userId的值
+        Long userId = (Long) session.getAttribute("userId");
+        if (userId == null) {
+            // 用户过期
+            return ResultVO.notLogin();
+        }
+        BannerPO bannerPO = bannerMapper.selectByPrimaryKey(bid);
         bannerMapper.deleteByPrimaryKey(bid);
+        logService.insertActionRecord(session,OpsRecordMenuConstant.F_ExternalFunction,OpsRecordMenuConstant.S_BannerConfig,
+                "删除banner\"%s\"",bannerPO==null?"":bannerPO.getName());
         return new ResultVO(1000);
     }
 
     @Override
-    public ResultVO upload(MultipartFile file) {
+    public ResultVO upload(HttpSession session,MultipartFile file) {
+        //从session中获取userId的值
+        Long userId = (Long) session.getAttribute("userId");
+        if (userId == null) {
+            // 用户过期
+            return ResultVO.notLogin();
+        }
         String picture = FileUtil.uploadFile(file, filePathConfig.getLocalBasePath(), filePathConfig.getProjectDir(), filePathConfig.getBannerDir(), "banner");
+        logService.insertActionRecord(session,OpsRecordMenuConstant.F_ExternalFunction,OpsRecordMenuConstant.S_BannerConfig,
+                "%s上传了banner图片",(String) session.getAttribute("username"));
         return new ResultVO<>(1000, filePathConfig.getPublicBasePath() + picture);
     }
 }
